fix: refactor cmd utils
